+++ source/jdbcDriverOOo/source/io/github/prrvchr/uno/sdbcx/Descriptor.java
@@ -76,13 +76,13 @@ public abstract class Descriptor
     @Override
     protected void registerProperties(Map<String, PropertyWrapper> properties) {
         short attribute;
+        PropertySetter setter;
+
         if (mReadonly) {
             attribute = PropertyAttribute.READONLY;
+            setter = null;
         } else {
             attribute = 0;
-        }
-        PropertySetter setter = null;
-        if (!mReadonly) {
             setter = value -> {
                 mName = (String) value;
             };
@@ -114,7 +114,6 @@ public abstract class Descriptor
         return ServiceInfo.supportsService(mServices, service);
     }
 
-
     // Method for internal use (no UNO method)
     protected String getName() {
         return mName;
@@ -127,6 +126,5 @@ public abstract class Descriptor
     protected boolean isCaseSensitive() {
         return mSensitive;
     }
-
 
 }